+++ src/Serilog.Enrichers.Sensitive/MaskOptions.cs
@@ -9,6 +9,7 @@ public class MaskOptions : IEquatable<MaskOptions>
     public int ShowFirst { get; set; } = NotSet;
     public int ShowLast { get; set; } = NotSet;
     public bool PreserveLength { get; set; } = true;
+    public bool WildcardMatch { get; set; }
 
     public bool Equals(MaskOptions? other)
     {
@@ -22,7 +23,12 @@ public class MaskOptions : IEquatable<MaskOptions>
             return true;
         }
 
-        return ShowFirst == other.ShowFirst && ShowLast == other.ShowLast && PreserveLength == other.PreserveLength;
+        if (other.GetType() != GetType())
+        {
+            return false;
+        }
+
+        return ShowFirst == other.ShowFirst && ShowLast == other.ShowLast && PreserveLength == other.PreserveLength && WildcardMatch == other.WildcardMatch;
     }
 
     public override bool Equals(object? obj)